requirements.txt
pymupdf
python-docx
openpyxl 
python-dotenv
openai
openai==1.12.0